% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_groups.R
\name{retirement_groups}
\alias{retirement_groups}
\title{Return valid OECD/EU retirement group labels}
\usage{
retirement_groups()
}
\value{
A character vector of valid group labels.
}
\description{
Return the valid group labels used in grouped maps and grouped time-series
plots.
}
\examples{
retirement_groups()

}
