% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_variables.R
\name{retirement_variables}
\alias{retirement_variables}
\title{Return available demographic variables for plotting}
\usage{
retirement_variables()
}
\value{
A character vector of demographic variable names.
}
\description{
Return the numeric demographic variables from \link{retirement} that can be used
in demographic plotting functions.
}
\details{
Raw retirement-age columns are excluded because retirement plotting is
handled through the derived \code{"retirement_age"} measure.
}
\examples{
retirement_variables()

}
