% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_year_measures.R
\name{retirement_year_measures}
\alias{retirement_year_measures}
\title{Return demographic measures expressed in years}
\usage{
retirement_year_measures()
}
\value{
A character vector of demographic variable names measured in years.
}
\description{
Return the demographic measures from \link{retirement} that are expressed in years
and are suitable for use in \code{\link[=plot_retirement_vs_demography]{plot_retirement_vs_demography()}}.
}
\examples{
retirement_year_measures()

}
