% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_var.R
\name{choose_var}
\alias{choose_var}
\title{Choose any variables from retirement data to use for analysis}
\usage{
choose_var(..., data = retirement)
}
\arguments{
\item{...}{Any variable chosen to keep in the final dataset}
}
\description{
If no variables are chosen, some general population statistics are kept
}
