% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retirement}
\alias{retirement}
\title{Retirement and demographic indicators}
\format{
A data frame with one row per country-year.
}
\usage{
retirement
}
\description{
A country-year dataset combining effective labour market exit age data from
the OECD with demographic indicators from the United Nations World
Population Prospects.
}
\keyword{datasets}
